Fix reset sequence order to keep right enc RB
Fix the order in which the reset recover sequence was performed
to prevent the encoder RB to be incorrectly set.
It is necessary to first recover the encoder RB value before
resetting the encoder (turning enc off and on again), because
the off/on operation generated spurious encoder values which
erased the latest encoder value saved at the 'SaveLastEncPos' PV,
which then was used for recovering the RB value.