After reconnect, store enc RB before disabling amp
When restoring controller parameters after a reconnect
by processing output records, store the encoder RB to be
used before changing the amplifier configuration,
since in some cases, disabling the amp can change the enc
RB value, causing a slightly different value for enc
RB to be set.

In the previous IOC version, the enc RB was saved AFTER
disabling the amplifier, which caused the error in the enc RB
recovered value, when UEIP (Use encoder) was disabled.